Fixed compiler warning regarding possible null pointer by doing a null check.
git-svn-id: file:///home/coren/project/backup_ho/trunk/Core@1019 b8de521a-fa13-0410-8113-fc7ffabbd59a

+++ src/java/de/hattrickorganizer/tools/backup/BackupHelper.java
@@ -62,11 +62,11 @@ public class BackupHelper {
 				 for (int i = 0; i < files.length; i++) {
 					if(i == 0 || (toDelete != null && toDelete.lastModified() > files[i].lastModified())){
 						toDelete = files[i];
-					}
-					 
+					}	 
 				}
-				 toDelete.delete();
-				 deleteOldFiles(dbDirectory);
+				if (toDelete != null)
+					toDelete.delete();
+				deleteOldFiles(dbDirectory);
 			 }
 		  }
